Interface moved to Properties, from UI
